[commtest] Move backup client to shared
Summary:
[ENG-5555](https://linear.app/comm/issue/ENG-5555) and [ENG-5554](https://linear.app/comm/issue/ENG-5554)

Move the backup client to shared so that it can be used on native. There are a few changes that I will annotate in inline comments. The total amount of moved logic is pretty small but if it's hard to review, let me know and I will split this into more diffs.

Test Plan:
- `cargo run` in backup and blob
- `yarn run-integration-tests backup`
- `yarn run-performance-tests blob`

Reviewers: bartek, varun, kamil

Reviewed By: bartek, varun

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10260