[lib] Convert getPotentialMemberItems into a hook
Summary:
This should be a no-op change in terms of the results that `getPotentialMemberItems` used to return. I'm simply converting it into a hook so I can call a hook that will construct a search index in the next diff, necessary to make it possible to prefix-search for users by their ENS names.

Addresses [[ https://linear.app/comm/issue/ENG-5402/convert-getpotentialmemberitems-to-a-hook | ENG-5402 ]]

Test Plan: Did some manual testing to confirm that there was no change in the search results returned (both by `console.log`'ing before and after, as well as searching for users on the app)

Reviewers: atul, ginsu, inka

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10229