[native][web] Cast InputState / BaseInputState when returning from selectors
Summary:
While working on the `unused-promise` Flow lint (in particular, "[web] Hide that InputState.sendTextMessage returns a Promise") I noticed that Flow was failing to typecheck the return value of these selectors.

Adding this cast fixed the issue, and things appear to be typechecked correctly now.

(@atul, this was the change that previously caused `minimallyEncoded` type errors for me, but no longer does after rebasing.)

Depends on D10167

Test Plan: Flow

Reviewers: atul

Subscribers: tomek, atul

Differential Revision: https://phab.comm.dev/D10168