[native] Remove async IIFE in CommunityDrawerContent effect
Summary:
This async IIFE was introduced in D7733. It's not clear to me why we need to `await` the `Promise` returned from `dispatchActionPromise` here, so I'm removing the `await` here.

If there's a reason for it, I can alternately add a `void` keyword preceding the async IIFE invocation to silence the `unused-promise` Flow lint.

Depends on D10173

Test Plan: Flow

Reviewers: tomek

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D10174