[lib][native][web][landing] Silence unused-promise using void keyword
Summary:
The Flow team recommends using the `void` keyword to silence the `unused-promise` Lint. See the "Fire-and-forget" part of [this blog post](https://medium.com/flow-type/flow-can-now-detect-unused-promises-b49341256640).

In this diff, I add the `void` keyword to a bunch of places in `native`, `web`, `landing`, and the parts of `lib` that are only used from `native`/`web`/`landing`. In these contexts it's safe to ignore promise rejections, because unlike Node.js there are no negative consequences to it. In contrast, we need to be more careful in `keyserver` (and potentially `desktop`?) because Node.js will crash when it encounters an unhandled promise rejection.

Depends on D10175

Test Plan: Flow

Reviewers: michal

Reviewed By: michal

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D10176