[keyserver] Ignore potential to crash on promise rejections in scripts
Summary:
Node.js runtimes such as `keyserver` can crash on unhandled promise rejections. In general, that's something we probably want to avoid.

However, in the context of scripts, we actually want an unhandled promise rejection to crash the script. In this diff, we add `void` declarations to silence the `unused-promise` Lint, and otherwise leave the behavior as-is.

Regarding the `void` keyword: the Flow team recommends using the `void` keyword to silence the `unused-promise` Lint. See the "Fire-and-forget" part of [this blog post](https://medium.com/flow-type/flow-can-now-detect-unused-promises-b49341256640).

Depends on D10176

Test Plan: Flow

Reviewers: michal

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D10177