[keyserver] Silence unused-promise Flow lint in backupDB
Summary:
The `async` IIFE in `backupDB` contains two `try-catch` blocks that guarantee it will never reject.

As such, we're safe to ignore the `unused-promise` Flow lint. Our primary reason for enabling that Lint is to catch unhandled promise rejections within `keyserver`, but there is no potential of that here.

Regarding the `void` keyword: the Flow team recommends using the `void` keyword to silence the `unused-promise` Lint. See the "Fire-and-forget" part of [this blog post](https://medium.com/flow-type/flow-can-now-detect-unused-promises-b49341256640).

Depends on D10177

Test Plan: Flow

Reviewers: michal

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D10178