[keyserver] Ignore potential to crash on promise rejections in keyserver preamble
Summary:
We use an `async` IIFE in lieu of top-level `await` (which Flow doesn't support) to initialize the keyserver.

In most cases we don't want to crash on promise rejections in `keyserver`. But if we encounter an error in the preamble, we should probably crash, since it indicates that the keyserver is not able to initialize itself.

This diff doesn't change any behavior. We're just using the `void` keyword to silence the `unused-promise` Lint, which is recommended by the Flow team. See the "Fire-and-forget" part of [this blog post](https://medium.com/flow-type/flow-can-now-detect-unused-promises-b49341256640).

Depends on D10178

Test Plan: Flow

Reviewers: michal

Subscribers: