[native] native_rust_library debug builds
Summary: we don't need to do release builds of the rust library when we're just doing a debug build of the iOS app

Test Plan: built the app for simulator and did archive builds a couple times in different orders, confirmed everything succeeded. checked the target folder to make sure we aren't doing release builds for debug builds of the app anymore.

Reviewers: bartek, ashoat

Subscribers: