[native] use build configuration to set socket addr
Summary:
Depends on D10186

Set the default socket address depending on the build configuration. We use staging for debug builds and prod for release builds. This mirrors the defaults on keyserver.

Test Plan:
the newly added unit test passed

after building with and without a JSON file present, and with/without `--release`, i made sure the correct `IDENTITY_SOCKET_ADDR` value was present

Reviewers: bartek, michal

Reviewed By: bartek, michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10187