[lib] Fix updatesCurrentAsOf returned from siweAuth
Summary: `siweAuth` returns `LogInResult` where `updatesCurrentAsOf` is of type `{+[keyserverID: string]: number}`. `response.serverTime` is a number. Flow didn't catch this because to it `response` is `any`. We use `ashoatKeyserverID` because this action was not yet refactored to work with multiple keyservers.

Test Plan: Checked that the types are now correct

Reviewers: kamil, michal, ginsu, ashoat

Reviewed By: ashoat

Subscribers: varun, ashoat, tomek

Differential Revision: https://phab.comm.dev/D10190