[lib] Update `convertRawThreadInfoToClientDBThreadInfo` to handle `RawThreadInfo`
Summary:
Basically modify function to pull out `minimallyEncoded` field before converting to `ClientDBThreadInfo` representation. This is fine because all of the columns in the SQLite database are compatible with the new "shape" of the data here.

Unfortunately, we can't do this as easily with `convertClientDBThreadInfoToRawThreadInfo` because it's used as part of migrations so we don't want to change the behavior of those migrations for users upgrading from an older version. Will have to create a "legacy" version of that function so migration functionality is retained.

---

Depends on D10197

Test Plan: CI/`flow`/etc

Reviewers: ashoat, ginsu, tomek, rohan

Differential Revision: https://phab.comm.dev/D10198