[keyserver] Add a function that synchronizes keyserver's invite links with blobs
Summary:
In this script we try to create blobs for links where a holder is missing or a holder doesn't point to an existing blob. This function will be used in a migration and in a script that periodically synchronizes data.

Depends on D10150

Test Plan: Run the function as a script and check if it creates blobs for all the links that don't have them.

Reviewers: michal, bartek, rohan

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D10205