identity] Add function to get current device list timestamp
Summary:

When doing a device list update, we need to make sure no concurrent update hapoened in meanwhile.
We can do it by storing last update time, then each update is a three-step transaction:
1. Get last update time
2. Do the update
3. Commit the update only if the last update time is the same as the one we got in step 1.

This function will return the current timestamp of the device list (a.k.a last update time).

Depends on D10215

Test Plan:

Manually created this attribute and checked that it is returned by the function.

Reviewers: michal, varun

Subscribers: