[lib] correctly skip using established socket when urlPrefix is overriden
Summary:
In D9667, I got some feedback about improving the solution for not using an established socket for calling endpoints whenever we override the `urlPrefix`. This diff addresses that feedback.

Depends on D9667

Test Plan:
Created this dummy params override object:

```
    const dummyOverride = {
      urlPrefixOverride: 'testing',
    };
```

And passed it to endpoint calls that would normally use an established socket and confirmed that when we overrode the url prefix with this dummy params override object we were not entering the condition in `callServerEndpoint` where we use an established socket

Reviewers: inka, michal, rohan, ashoat

Reviewed By: inka, ashoat

Subscribers: ashoat, tomek, wyilio

Differential Revision: https://phab.comm.dev/D9673