[lib] Introduce `deprecatedConvertClientDBThreadInfoToRawThreadInfo`
Summary:
Introduce `deprecatedConvertClientDBThreadInfoToRawThreadInfo` which turns a ClientDB representation of `RawThreadInfo` into `LegacyRawThreadInfo`.

The reason for this function, as described in D10198, is to maintain correctness of existing Redux migrations.

(As an aside, Redux migrations would be a great thing to write unit tests for because often our assumptions about "current state of the codebase" when modifying helper functions don't apply to migrations and can break them. I just happened to notice `convertClientDBThreadInfoToRawThreadInfo` when looking at creating a new Redux migration and realized this would break old migrations.)

---

Depends on D10209

Test Plan:
NA/CI/`flow`

NOTE: Disabled two unit tests for now, will re-enable after updating test data. This is definitely a flow I'm planning on testing so put in `TODO`s for now... thought happy to create Linear tasks if that'd be preferred.

Reviewers: ashoat, ginsu, tomek, rohan

Reviewed By: rohan

Differential Revision: https://phab.comm.dev/D10210