[web] introduce subheader prop to modal component
Summary:
This diff introduces a `subheader` prop to the modal component. Sometimes we want to render an element in between the header and the modal children content, but we want this element to not follow the spacing rules for the modal children. These elements can be things like banners, tabs, etc. For example:

{F916853}

The `subheader` prop will extend the entire width of the modal, perfect for elements like banners + tabs.

Linear task: https://linear.app/comm/issue/ENG-5943/extendmodify-the-modal-props-api-to-follow-new-modal-designs

Depends on D10232

Test Plan:
Here is an example of using the `subheader` prop with the community creation modal

{F916887}

Reviewers: atul, rohan, kamil

Reviewed By: rohan

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10234