[web] introduce subheader prop to modal component
Summary: This diff introduces a `subheader` prop to the modal component. Sometimes we want to render an element in between the header and the modal children content, but we want this element to not follow the spacing rules for the modal children.

Test Plan: Here is an example of using the `subheader` prop with the community creation modal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10234