[web] match the veritcal padding for modals with no button containers
Summary:
Some modals will not have a CTA/button section. If that is the case the padding of the modal will be top: 32px, horizontal: 32px and bottom: 24px. This is based off the rules below:
{F916981}

To make sure the vertical padding for modals is equal to each other when there is no button section, I added an empty div with a height of 8px to act as on offset so that if there is no button section then the modal will just have a 32px uniform padding across all of its sides

Depends on D10236

Test Plan:
Please see screenshot of my inspector showing the correct dimensions below for a modal with no CTA/button section

top has padding of 32px:
{F916988}

horizontal sides have padding of 32px + bottom of 24px
{F916991}

bottom offset of 8px (24+8 = 32)
{F916992}

Reviewers: atul, rohan, kamil

Reviewed By: rohan

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10237