[native-rust-library] Backup service config
Summary:
[ENG-5346](https://linear.app/comm/issue/ENG-5345)

This diff introduces the config for backup URL and refactors the existing identity service config code so that it can be reused between identity and backup.

Test Plan:
Added `assert_eq!(BACKUP_SOCKET_ADDR, "");` in the test to get the current value of `BACKUP_SOCKET_ADDR` in the panic message. Then I've run these command and checked the value of `BACKUP_SOCKET_ADDR`:
- `cargo test --release`, no file -> `backup.commtechnologies.org`
- `cargo test --release`, file -> value in the file
- `cargo test`, file -> value in the file
- `cargo test`, no file -> `backup.staging.commtechnologies.org`

Reviewers: varun, bartek, kamil

Reviewed By: varun, bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10261