[native-rust-library] Add backup client dependency
Summary:
[ENG-5556](https://linear.app/comm/issue/ENG-5556)
Add the backup client to the native_rust library. We activate different features depending on a platform - on iOS we can use the native tls implementation but on Android the native tls reqwest feature pulls in openssl (beause it's linux) and it conflicts with our openssl during linking process. Because of that we replace it with a pure-rust implementation `rustls-tls`.

Depends on D10261

Test Plan:
Test that the app compiles on an physical Android and iOS devices.

Reviewers: bartek, varun, kamil

Subscribers: