[keyserver] Update `rawThreadInfoFromServerThreadInfo` to support `minimallyEncodePermissions` option
Summary:
This is the diff that **finally** allows `keyserver` to send the new "minimally encoded" permissions to new clients.

We introduce `minimallyEncodePermissions` option, which for now will be set to `NATIVE` `FUTURE_CODE_VERSION`. Once we have a version of client that fully supports minimally encoded permissions, we will update this setting.

---

Depends on D10285

Test Plan:
1. Will set breakpoint in `keyserver` and ensure that `rawThreadInfos` are constructed as expected.

{F924039}

2. Will check Redux dev tools on corresponding client to ensure that minimally encoded permissions appear as expected.
3. Will play around with the `native` app and make sure that permissions checks continue to behave as expected.

Reviewers: ashoat, ginsu, tomek, rohan

Reviewed By: rohan

Differential Revision: https://phab.comm.dev/D10286