[lib][web][native] Simplify logout call
Summary:
Now that actions are wrapped in hooks we can select data in them. This simplfies the calling code and reduces repetition.

Test Plan:
Checked that it is possible to log out and back in with no errors showing up. Checked that `invalidSessionDowngrade` works correctly (by loggin out and back in before `logoutActionTypes.success` was dispatched)

Reviewers:
tomek, michal, varun
Subscribers: