[lib] Call logout when connectionIssue is set
Summary:
issue: https://linear.app/comm/issue/ENG-5914/stop-calling-logout-from-socket
See [[ https://linear.app/comm/issue/ENG-5913/check-all-places-that-select-all-keyservers-from-keyserverstore#comment-3cbd4a93 | this comment ]]. We want to stop calling logout from the Socket, and instead set `conneciotnIssue`. But for the time baing we will log the user out from `KeyserverConnectionHandler` if `connectionIssue` was set. This will be updated in the future to instead retry the keyserver from time to time.

Test Plan: Dispatched the `setConnectionIssueActionTypes`, checked that the user is logged out.

Reviewers: michal, kamil, ginsu, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D10331