[keyserver] Move `threadStoreValidator` to `redux-state-responders`
Summary:
This diff is a follow-up diff to D10294.

Initially when I tried updating `threadStoreValidator` I had some issues with Jest... and as before it was a circular dependency issue.

Since `threadStoreValidator` is only being used by `initialReduxStateValidator`, I moved it right beside (above) it.

---

Depends on D10294

Test Plan: CI/`flow`/etc. Will become clear whether or not this validator works when we get to `web`.

Reviewers: ashoat, ginsu, tomek, rohan

Differential Revision: https://phab.comm.dev/D10295