[lib/native] Show resolved ENS name in user profile
Summary: Noticed this while working on my ENS work, but the user profile doesn't show the resolved ENS name on both `web` and `native`. Thought I'd make this quick change, but @ginsu feel free to let me know if this was either intentional / not the right way to handle this.

Test Plan:
Confirmed the ENS name shows if available now

{F919196}

{F919197}

Reviewers: ginsu, atul

Reviewed By: ginsu

Subscribers: ashoat, tomek, ginsu

Differential Revision: https://phab.comm.dev/D10259