[lib] Do not provide overrides when we already know a keyserver
Summary:
If we know a keyserver, we have a cookie for it. If we use an override, we don't send that cookie, which means that from the keyserver's point of view, we're a new user. The solution is to provide an override only if we don't already know a keyserver.

Depends on D10290

Test Plan: Provide an override to a local keyserver when a user is logged in, click a link and check if the login screen doesn't appear.

Reviewers: michal, bartek, rohan, inka

Reviewed By: rohan

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D10307