[keyserver] Modify link verification so that it works with users that aren't logged in to a keyserver
Summary:
Initially we were assuming that all the users that click an invite link are logged in to a keyserver. In multi-keyserver world it can happen that a user is logged in to Identity service, but isn't to a keyserver.

Before logging the user in, we need to check if a link is valid - that requires modifying the verification endpoint so that it no longer assumes that a user is logged in. This assumption was used only to determine if a user is a member of a thread - which is obviously false for a user that isn't logged in.

Depends on D10307

Test Plan: Bring back a bug from D10307 which allows easily checking what happens when a user isn't logged in. Check if the endpoint works for users that are and aren't logged in.

Reviewers: michal, bartek, rohan, inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D10308