[lib] Allow overriding a keyserver in a function returned from useKeyserverCall hook
Summary:
Modifying the keyserver by setting a parameter of a hook appeared to be really complicated when we don't know the keyserver when rendering a component. This diff allows overriding when calling a function.

Depends on D10308

Test Plan: Check if calling link verification works with default and overridden keyserver.

Reviewers: michal, bartek, rohan, inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D10311