[CommRustModule] update `getOutboundKeys` to return all devices
Summary:
To create a session with all other devices of a given user we don't know the IDs of specific devices so we need to get all of them.
I think it's more convenient to update this method (later we can filter if we want a specific ID) rather than introducing a new JSI method.

Depends on D10372

Test Plan:
1. Login to staging Identity (`loginPasswordUser`) on two devices using the same account.
2. Call this method and check if the result is correct (two devices).

Reviewers: marcin, varun, michal