[lib] Update `convertClientDBThreadInfoToRawThreadInfo` to validate `RoleInfo`s
Summary:
Similarly to D10298, this addresses what we learned when testing D10297 that it's important that we validate the stringified JSON that we retrieve from `ClientDBThreadInfo`. Specifically, it's possible for `ClientDBThreadInfo` to have either stringified `MinimallyEncodedRoleInfo`s or `LegacyEncodedRoleInfo`s stored in the `.roles` field, and we should A. valdidate that the data is correctly formed B. determine which format the data is in.

---

Depends on D10298

Test Plan: Will be tested along with changes in D10297 and D10298

Reviewers: ashoat, ginsu, tomek, rohan

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D10313